refactor: extract image commands into ai-image.ts (-1,050 lines from ai.ts)
Extracts 11 image commands + executeThumbnailBestFrame into ai-image.ts:

ai-image.ts (~1,100 lines):
  Commands: image, thumbnail, background,
            sd, sd-upscale, sd-remove-bg, sd-img2img, sd-replace, sd-outpaint,
            gemini, gemini-edit
  + executeThumbnailBestFrame (moved from execute section, used by thumbnail cmd)
  registerImageCommands(cmd) — called from ai.ts

ai.ts:
  Re-exports executeThumbnailBestFrame for backward compat
  Shrinks from 11,863 → 10,813 lines (-1,050 net)

All 232 tests pass.
